Move config and default to relaxed validation mode if not provided
